Reorganize combinators into dedicated folder
