+++ src/AggregateFunctions/combinators/AggregateFunctionArray.cpp
@@ -1,7 +1,7 @@
-#include <AggregateFunctions/AggregateFunctionArray.h>
-#include <AggregateFunctions/AggregateFunctionCombinatorFactory.h>
-#include <Common/typeid_cast.h>
+#include "AggregateFunctionArray.h"
+#include "AggregateFunctionCombinatorFactory.h"
 
+#include <Common/typeid_cast.h>
 
 namespace DB
 {
